+++ tests/unit/ProctorioRequestHandlerTest.php
@@ -23,7 +23,9 @@ declare(strict_types=1);
 namespace oat\Proctorio\tests\unit;
 
 use GuzzleHttp\ClientInterface;
+use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Psr7\Request;
+use oat\Proctorio\Exception\InvalidProctorioResponseException;
 use oat\Proctorio\ProctorioRequestHandler;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
@@ -50,9 +52,21 @@ class ProctorioRequestHandlerTest extends TestCase
         );
     }
 
-    public function testBuildRequest(): void
+    public function testExecuteRequestSuccessfully(): void
     {
         $payload = '';
+
+        $this->responseMock
+            ->method('getBody')
+            ->willReturn(
+                json_encode(
+                    [
+                        'url1',
+                        'url2',
+                    ]
+                )
+            );
+
         $this->clientMock
             ->expects($this->once())
             ->method('send')
@@ -68,11 +82,123 @@ class ProctorioRequestHandlerTest extends TestCase
                             CURLOPT_SSLVERSION => CURL_SSLVERSION_TLSv1_2,
                         ],
                     ],
-                    ''
+                    $payload
                 )
             )
             ->willReturn($this->responseMock);
 
-        $this->subject->execute($payload);
+        $proctorioResponse = $this->subject->execute($payload);
+
+        $this->assertSame('url1', $proctorioResponse->getTestTakerUrl());
+        $this->assertSame('url2', $proctorioResponse->getTestReviewerUrl());
+    }
+
+    public function testExecuteRequestReturnsInvalidResponse(): void
+    {
+        $this->responseMock
+            ->method('getBody')
+            ->willReturn(
+                json_encode(
+                    [
+                        'url1'
+                    ]
+                )
+            );
+
+        $this->clientMock
+            ->expects($this->once())
+            ->method('send')
+            ->willReturn($this->responseMock);
+
+        $this->expectException(InvalidProctorioResponseException::class);
+        $this->expectExceptionMessage('Invalid Proctorio response');
+
+        $this->subject->execute('');
+    }
+
+    public function testExecuteRequestThrowsGuzzleException(): void
+    {
+        $this->responseMock
+            ->method('getBody')
+            ->willReturn(
+                json_encode(
+                    [
+                        'url1'
+                    ]
+                )
+            );
+
+        $this->clientMock
+            ->expects($this->once())
+            ->method('send')
+            ->willThrowException($this->createMock(GuzzleException::class));
+
+        $this->expectException(InvalidProctorioResponseException::class);
+        $this->expectExceptionMessage('Invalid Proctorio response');
+
+        $this->subject->execute('');
+    }
+
+    /**
+     * @dataProvider invalidResponseCodeProvider
+     */
+    public function testExecuteRequestReturnsInvalidResponseCode(int $responseCode, string $expectedMessage): void
+    {
+        $responseData = json_encode([$responseCode, $expectedMessage]);
+
+        $this->responseMock
+            ->method('getBody')
+            ->willReturn($responseData);
+
+        $this->clientMock
+            ->expects($this->once())
+            ->method('send')
+            ->willReturn($this->responseMock);
+
+        try {
+            $this->subject->execute('');
+        } catch (InvalidProctorioResponseException $exception) {
+            $this->assertTrue($exception instanceof InvalidProctorioResponseException);
+            $this->assertSame($expectedMessage, $exception->getMessage());
+            $this->assertSame($responseData, $exception->getResponseData());
+        }
+    }
+
+    public function invalidResponseCodeProvider(): array
+    {
+        return [
+            [
+                2653,
+                'Missing required parameters',
+            ],
+            [
+                2654,
+                'Invalid parameter',
+            ],
+            [
+                2655,
+                'Incorrect consumer key',
+            ],
+            [
+                2656,
+                'Signature is invalid'
+            ],
+            [
+                2657,
+                'The used timestamp is out of range'
+            ],
+            [
+                2658,
+                'Invalid exam tag ID'
+            ],
+            [
+                2659,
+                'Invalid settings'
+            ],
+            [
+                2660,
+                'Unknown'
+            ],
+        ];
     }
 }