+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 CHANGELOG
 =========
 
-0.0.0
+2.0.0
 -----
+- Added validator for `ProctorioConfig` payload.
+- Validate response from Proctorio inside library.
+- Removed exposal of Guzzle exception to library's clients.
+- Returning self-contained `ProctorioResponse` object instead of raw array.
+
+1.0.0
+-----
+- First Proctorio stable release.
+- Possibility to request testTaker and testReviewer URLs from Proctorio.